+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+### Changed
+### Fixed
+
+## [3.0.0] - 2024-08-19
+### Added
 - Query parameter operator can be "ANY" to find rows where an array column contains the given value
 ### Changed
+- Upgrade PgQuery gem
 ### Fixed
 - Prevent exception when parsing a CASE that returns the value of a column instead of a const